Renames TELEMETRY setting to ANALYTICS.
fixes: #3417
