+++ CHANGES/3213.bugfix
@@ -1,2 +1,2 @@
 Fixed bug where installations with at least one pre-release plugin installed were posting to the
-production telemetry site instead of the developer telemetry site.
+production analytics site instead of the developer analytics site.
