+++ CHANGES/3417.deprecation
@@ -1,0 +1,1 @@
+``TELEMETRY` setting was deprecated in favor of ``ANALYTICS``.
