+++ CHANGES/3417.feature
@@ -1,0 +1,2 @@
+The new ``ANALYTICS`` setting replaced the ``TELEMETRY`` setting to avoid confusion with
+application metrics that will be exposed using OpenTelemetry.
